+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/agent/protocol_generated/ConfigFeaturesParams.kt
@@ -1,0 +1,9 @@
+@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
+package com.sourcegraph.cody.agent.protocol_generated;
+
+data class ConfigFeaturesParams(
+  val chat: Boolean,
+  val attribution: Boolean,
+  val serverSentModels: Boolean,
+)
+