+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/ContextItemSource.kt
@@ -1,5 +1,0 @@
-@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated;
-
-typealias ContextItemSource = String // One of: embeddings, user, editor, search, initial, unified, selection, terminal, uri, history
-