+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/TextDocumentShowOptionsParams.kt
@@ -1,9 +1,0 @@
-@file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
-package com.sourcegraph.cody.protocol_generated;
-
-data class TextDocumentShowOptionsParams(
-  val preserveFocus: Boolean? = null,
-  val preview: Boolean? = null,
-  val selection: Range? = null,
-)
-