+++ vscode/src/chat/chat-view/ChatsController.ts
@@ -28,6 +28,7 @@ import { getConfiguration } from '../../configuration'
 import type { EnterpriseContextFactory } from '../../context/enterprise-context-factory'
 import type { ContextRankingController } from '../../local-context/context-ranking'
 import type { AuthProvider } from '../../services/AuthProvider'
+import type { ContextAPIClient } from '../context/contextAPIClient'
 import {
     ChatController,
     type ChatSession,
@@ -69,7 +70,8 @@ export class ChatsController implements vscode.Disposable {
         private readonly localEmbeddings: LocalEmbeddingsController | null,
         private readonly contextRanking: ContextRankingController | null,
         private readonly symf: SymfRunner | null,
-        private readonly guardrails: Guardrails
+        private readonly guardrails: Guardrails,
+        private readonly contextAPIClient: ContextAPIClient | null
     ) {
         logDebug('ChatsController:constructor', 'init')
         this.panel = this.createChatController()
@@ -147,9 +149,6 @@ export class ChatsController implements vscode.Disposable {
             vscode.commands.registerCommand('cody.action.chat', args =>
                 this.submitChatInNewEditor(args)
             ),
-            vscode.commands.registerCommand('cody.chat.focusView', () =>
-                vscode.commands.executeCommand('cody.chat.focus')
-            ),
             vscode.commands.registerCommand('cody.chat.signIn', () =>
                 vscode.commands.executeCommand('cody.chat.focus')
             ),
@@ -157,6 +156,16 @@ export class ChatsController implements vscode.Disposable {
                 await this.panel.clearAndRestartSession()
                 await vscode.commands.executeCommand('cody.chat.focus')
             }),
+            vscode.commands.registerCommand(
+                'cody.chat.toggle',
+                async (ops: { editorFocus: boolean }) => {
+                    if (ops.editorFocus) {
+                        await vscode.commands.executeCommand('cody.chat.focus')
+                    } else {
+                        await vscode.commands.executeCommand('workbench.action.focusActiveEditorGroup')
+                    }
+                }
+            ),
             vscode.commands.registerCommand('cody.chat.newEditorPanel', () =>
                 this.getOrCreateEditorChatController()
             ),
@@ -201,6 +210,9 @@ export class ChatsController implements vscode.Disposable {
     private async sendEditorContextToChat(uri?: URI): Promise<void> {
         telemetryRecorder.recordEvent('cody.addChatContext', 'clicked')
         const provider = await this.getActiveChatController()
+        if (provider === this.panel) {
+            await vscode.commands.executeCommand('cody.chat.focus')
+        }
         await provider.handleGetUserEditorContext(uri)
     }
 
@@ -431,6 +443,7 @@ export class ChatsController implements vscode.Disposable {
             models,
             guardrails: this.guardrails,
             startTokenReceiver: this.options.startTokenReceiver,
+            contextAPIClient: this.contextAPIClient,
         })
     }