+++ web/lib/agent/agent.client.ts
@@ -28,6 +28,7 @@ interface AgentClientOptions {
     serverEndpoint: string
     accessToken: string
     workspaceRootUri: string
+    customHeaders?: Record<string, string>
     debug?: boolean
     trace?: boolean
 }
@@ -36,6 +37,7 @@ export async function createAgentClient({
     serverEndpoint,
     accessToken,
     workspaceRootUri,
+    customHeaders,
     debug = true,
     trace = false,
 }: AgentClientOptions): Promise<AgentClient> {
@@ -74,10 +76,13 @@ export async function createAgentClient({
         name: 'web',
         version: '0.0.1',
         workspaceRootUri,
+        capabilities: {
+            completions: 'none',
+        },
         extensionConfiguration: {
             accessToken,
             serverEndpoint,
-            customHeaders: {},
+            customHeaders: customHeaders ?? {},
             customConfiguration: {
                 'cody.experimental.noodle': true,
                 'cody.autocomplete.enabled': false,